Reduce risk of settings loss by abrupt process exit
Instead of overwriting the old file with the new file, the new file is
written to disk with a .writing extension. Then when writing is complete
the old file is deleted and the new files' extension is fixed.

Hopefully this will make sure no settings are lost when the process is
halted mid-write.